requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3
pydantic==2.10.0
python-dotenv==1.0.0